Fix empty server nonce (#30)
Deriving from the protocol description r2 is supposed to be a "random 256-bit server nonce." Instead of generating and writing it into r2 it's written into r which is then never used again. r2 remains an empty byte-array of length 32. I suppose this was a typo. This should fix it.